core: move Context and Value to top of the crate
I would like to remove the `core` module because it's a categorization
that doesn't really make sense anymore IMHO. Value and Context should
live in the crate root. The Node types I will replace in a later PR.
