+++ jets-bench/benches/elements/data_structures.rs
@@ -6,7 +6,7 @@ pub use bitcoin_hashes::sha256;
 use bitcoin_hashes::{hex::FromHex, Hash};
 use elements::Txid;
 use rand::{thread_rng, RngCore};
-use simplicity::{bitcoin, core::Value, elements, types::Type, BitIter, Error};
+use simplicity::{bitcoin, elements, types::Type, BitIter, Error, Value};
 
 /// Engine to compute SHA256 hash function.
 /// We can't use bitcoin_hashes::sha256::HashEngine because it does not accept
