+++ src/core/redeem.rs
@@ -12,7 +12,6 @@
 // If not, see <http://creativecommons.org/publicdomain/zero/1.0/>.
 //
 
-use crate::core::Value;
 use crate::dag::{DagLike, FullSharing, InternalSharing, NoSharing, PostOrderIter};
 use crate::decode::WitnessDecoder;
 use crate::jet::Jet;
@@ -21,7 +20,7 @@ use crate::merkle::cmr::Cmr;
 use crate::merkle::imr::Imr;
 use crate::types::{self, arrow::FinalArrow};
 use crate::{encode, Error};
-use crate::{BitIter, BitWriter};
+use crate::{BitIter, BitWriter, Value};
 use std::rc::Rc;
 use std::{fmt, io};
 
@@ -209,7 +208,6 @@ mod tests {
 
     use super::*;
 
-    use crate::core::Value;
     use crate::jet::Core;
 
     #[test]
