+++ src/lib.rs
@@ -36,6 +36,7 @@ mod macros;
 mod analysis;
 pub mod bit_encoding;
 pub mod bit_machine;
+mod context;
 pub mod core;
 pub mod dag;
 pub mod jet;
@@ -46,6 +47,7 @@ pub mod types;
 // #[cfg(test)]
 // mod test_progs;
 mod util;
+mod value;
 
 use bit_encoding::bititer; // FIXME used by autogenerator jet code
 pub use bit_encoding::bititer::EarlyEndOfStreamError;
@@ -59,12 +61,14 @@ pub use bit_encoding::BitWriter;
 pub use crate::policy::Policy;
 
 pub use crate::bit_machine::exec;
+pub use crate::context::Context;
 pub use crate::core::commit::CommitNodeInner;
 pub use crate::core::redeem::RedeemNodeInner;
-pub use crate::core::{CommitNode, Context, RedeemNode};
+pub use crate::core::{CommitNode, RedeemNode};
 pub use crate::decode::{decode_program, WitnessDecoder};
 pub use crate::encode::{encode_natural, encode_value, encode_witness};
 pub use crate::merkle::{amr::Amr, cmr::Cmr, imr::Imr, tmr::Tmr};
+pub use crate::value::Value;
 pub use simplicity_sys as ffi;
 use std::fmt;
