+++ src/merkle/cmr.rs
@@ -13,11 +13,10 @@
 //
 
 use crate::core::commit::CommitNodeInner;
-use crate::core::Value;
 use crate::impl_midstate_wrapper;
 use crate::jet::Jet;
-use crate::merkle::tmr::Tmr;
 use crate::merkle::{CommitMerkleRoot, MerkleRoot};
+use crate::{Tmr, Value};
 use bitcoin_hashes::sha256::Midstate;
 
 /// Commitment Merkle root
